- started cleaning up the import of low level functions - shifted low level functions from the component or pf files to the new folder "utils" - shifted the file pf.result_extraction to utils
