+++ src/pandapipes/__init__.py
@@ -13,6 +13,8 @@ import os
 pd.options.mode.chained_assignment = None  # default='warn'
 pp_dir = os.path.dirname(os.path.realpath(__file__))
 
+
+from pandapipes.utils.internals import *
 from pandapipes.properties.fluids import *
 from pandapipes.component_models.component_registry import *
 from pandapipes.create import *
