+++ src/pandapipes/component_models/_branch_models.py
@@ -8,7 +8,7 @@ from numpy import array
 from pandapipes.component_models._base_component import Component
 from pandapipes.component_models.junction_component import Junction
 from pandapipes.idx_branch import MDOTINIT, branch_cols, TEXT
-from pandapipes.pf.pipeflow_setup import get_table_number, get_lookup, get_net_option
+from pandapipes.utils.internals import get_net_option, get_table_number, get_lookup
 
 
 class BranchComponent(Component):
