+++ src/pandapipes/component_models/_circulation_pump.py
@@ -8,8 +8,8 @@ from pandapipes.component_models._branch_element_models import BranchElementComp
 from pandapipes.component_models.component_toolbox import set_fixed_node_entries, standard_branch_wo_internals_result_lookup
 from pandapipes.idx_branch import D, AREA, BRANCH_TYPE, CIRC, LOAD_VEC_BRANCHES_T, TO_NODE
 from pandapipes.idx_node import MDOTSLACKINIT, VAR_MASS_SLACK, JAC_DERIV_MSL
-from pandapipes.pf.pipeflow_setup import get_fluid
-from pandapipes.pf.result_extraction import extract_branch_results_without_internals
+from pandapipes.properties.fluids import get_fluid
+from pandapipes.utils.result_extraction import extract_branch_results_without_internals
 
 
 class CirculationPump(BranchElementComponent):
