+++ src/pandapipes/component_models/_const_flow_models.py
@@ -2,12 +2,11 @@
 # and Energy System Technology (IEE), Kassel, and University of Kassel. All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
 
-import numpy as np
-from numpy import dtype
+from numpy import dtype, nan_to_num, isin
+
 from pandapipes.component_models._node_element_models import NodeElementComponent
 from pandapipes.idx_node import LOAD, ELEMENT_IDX
-from pandapipes.pf.internals_toolbox import _sum_by_group
-from pandapipes.pf.pipeflow_setup import get_lookup, get_net_option
+from pandapipes.utils.internals import get_net_option, get_lookup, _sum_by_group
 
 
 class ConstFlow(NodeElementComponent):
@@ -23,7 +22,7 @@ class ConstFlow(NodeElementComponent):
         """
         loads = net[self.table_name]
         helper = loads.in_service.values * loads.scaling.values * self.sign
-        mf = np.nan_to_num(loads.mdot_kg_per_s.values)
+        mf = nan_to_num(loads.mdot_kg_per_s.values)
         mass_flow_loads = mf * helper
         juncts, loads_sum = _sum_by_group(get_net_option(net, "use_numba"), loads.junction.values,
                                           mass_flow_loads)
@@ -54,7 +53,7 @@ class ConstFlow(NodeElementComponent):
         fj, tj = get_lookup(net, "node", "from_to")[self.connected_node_type.table_name]
         junct_pit = net["_pit"]["node"][fj:tj, :]
         nodes_connected_hyd = get_lookup(net, "node", "active_hydraulics")[fj:tj]
-        is_juncts = np.isin(loads.junction.values, junct_pit[nodes_connected_hyd, ELEMENT_IDX])
+        is_juncts = isin(loads.junction.values, junct_pit[nodes_connected_hyd, ELEMENT_IDX])
 
         is_calc = is_loads & is_juncts
         res_table["mdot_kg_per_s"].values[is_calc] = loads.mdot_kg_per_s.values[is_calc] \
