+++ src/pandapipes/component_models/component_toolbox.py
@@ -3,100 +3,12 @@
 # Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
 
 import numpy as np
-import pandas as pd
 
-from pandapipes import get_fluid
-from pandapipes.constants import NORMAL_PRESSURE, TEMP_GRADIENT_KPM, AVG_TEMPERATURE_K, \
-    HEIGHT_EXPONENT
+from pandapipes.properties.fluids import get_fluid
 from pandapipes.idx_branch import LOAD_VEC_NODES_FROM, LOAD_VEC_NODES_TO, FROM_NODE, TO_NODE
 from pandapipes.idx_node import (EXT_GRID_OCCURENCE, EXT_GRID_OCCURENCE_T,
                                  PINIT, NODE_TYPE, P, TINIT, NODE_TYPE_T, T, LOAD)
-from pandapipes.pf.pipeflow_setup import get_net_option, get_lookup
-from pandapipes.pf.internals_toolbox import _sum_by_group
-
-
-def p_correction_height_air(height):
-    """
-
-    :param height:
-    :type height:
-    :return:
-    :rtype:
-    """
-    return NORMAL_PRESSURE * np.power(1 - height * TEMP_GRADIENT_KPM / AVG_TEMPERATURE_K,
-                                      HEIGHT_EXPONENT)
-
-
-def vinterp(min_vals, max_vals, lengths):
-    """
-
-    :param min_vals:
-    :type min_vals:
-    :param max_vals:
-    :type max_vals:
-    :param lengths: lengths for each range (same length as starts)
-    :type lengths: numpy.array
-    :return:
-    :rtype:
-    """
-    intervals = (max_vals - min_vals) / (lengths + 1)
-    steps = np.repeat(intervals, lengths)
-    counter = np.arange(lengths.sum()) - np.repeat(lengths.cumsum() - lengths, lengths) + 1
-    return np.repeat(min_vals, lengths) + steps * counter
-
-
-def vrange(starts, lengths):
-    """
-    Create concatenated ranges of integers for multiple start/length
-
-    :param starts: starts for each range
-    :type starts: numpy.array
-    :param lengths: lengths for each range (same length as starts)
-    :type lengths: numpy.array
-    :return: cat_range - concatenated ranges
-    :rtype: numpy.array
-
-    :Example:
-    >>> starts = np.array([1, 3, 4, 6])
-    >>> lengths = np.array([0, 2, 3, 0])
-    >>> print vrange(starts, lengths)
-    """
-    # Repeat start position index length times and concatenate
-    starting_array = np.repeat(starts, lengths)
-    # Create group counter that resets for each start/length
-    length_ranges = np.arange(lengths.sum()) - np.repeat(lengths.cumsum() - lengths, lengths)
-    # Add group counter to group specific starts
-    return starting_array + length_ranges
-
-
-def init_results_element(net, element, output, all_float):
-    """
-
-    :param net: The pandapipes network
-    :type net: pandapipesNet
-    :param element:
-    :type element:
-    :param output:
-    :type output:
-    :param all_float:
-    :type all_float:
-    :return: No Output.
-    """
-    res_element = "res_" + element
-    if all_float:
-        net[res_element] = pd.DataFrame(np.nan, columns=output, index=net[element].index,
-                                        dtype=np.float64)
-    else:
-        net[res_element] = pd.DataFrame(np.zeros(0, dtype=output), index=[])
-        net[res_element] = pd.DataFrame(np.nan, index=net[element].index,
-                                        columns=net[res_element].columns)
-
-
-def set_entry_check_repeat(pit, column, entry, repeat_number, repeated=True):
-    if repeated:
-        pit[:, column] = np.repeat(entry, repeat_number)
-    else:
-        pit[:, column] = entry
+from pandapipes.utils.internals import get_net_option, get_lookup, _sum_by_group
 
 
 def set_fixed_node_entries(net, node_pit, junctions, types, values, node_comp, mode):
