+++ src/pandapipes/component_models/flow_control_component.py
@@ -4,12 +4,12 @@
 
 from numpy import dtype, bool_, float64, zeros
 
-from pandapipes.component_models import BranchElementComponent
-from pandapipes.properties import get_fluid
-from pandapipes.component_models.component_toolbox import \
-    standard_branch_wo_internals_result_lookup, get_component_array
+from pandapipes.component_models._branch_element_models import BranchElementComponent
+from pandapipes.component_models.component_toolbox import standard_branch_wo_internals_result_lookup, \
+    get_component_array
 from pandapipes.idx_branch import JAC_DERIV_DP, JAC_DERIV_DP1, JAC_DERIV_DM, MDOTINIT, LOAD_VEC_BRANCHES
-from pandapipes.pf.result_extraction import extract_branch_results_without_internals
+from pandapipes.properties.fluids import get_fluid
+from pandapipes.utils.result_extraction import extract_branch_results_without_internals
 
 
 class FlowControlComponent(BranchElementComponent):
