+++ src/pandapipes/component_models/mass_storage_component.py
@@ -2,9 +2,9 @@
 # and Energy System Technology (IEE), Kassel, and University of Kassel. All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
 
-from pandapipes.component_models import ConstFlow
 from numpy import dtype
 
+from pandapipes.component_models._const_flow_models import ConstFlow
 
 class MassStorage(ConstFlow):
     """
