+++ src/pandapipes/component_models/valve_component.py
@@ -4,11 +4,11 @@
 
 from numpy import dtype, pi
 
-from pandapipes.component_models.component_toolbox import standard_branch_wo_internals_result_lookup
 from pandapipes.component_models._branch_element_models import BranchElementComponent
+from pandapipes.component_models.component_toolbox import standard_branch_wo_internals_result_lookup
 from pandapipes.idx_branch import D, AREA, LOSS_COEFFICIENT as LC
-from pandapipes.pf.result_extraction import extract_branch_results_without_internals
 from pandapipes.properties.fluids import get_fluid
+from pandapipes.utils.result_extraction import extract_branch_results_without_internals
 
 
 class Valve(BranchElementComponent):
