+++ src/pandapipes/converter/stanet/preparing_steps.py
@@ -10,6 +10,7 @@ import pandas as pd
 import pandapipes
 from pandapipes.converter.stanet.table_creation import CLIENT_TYPES_OF_NODES
 from pandapipes.properties.fluids import FluidPropertySutherland, _add_fluid_to_net
+from pandapipes.utils.internals import set_user_pf_options
 
 try:
     from shapely.geometry import Point, LineString
@@ -183,7 +184,7 @@ def get_net_params(net, stored_data):
     net_params["t0_sutherland"] = net_data.at[0, "T0"]
     net_params["calculate_temp"] = str(net_data.at[0, "TEMPCALC"]) == "J"
     pp_calc_mode = "sequential" if net_params["calculate_temp"] else "hydraulics"
-    pandapipes.set_user_pf_options(net, mode=pp_calc_mode)
+    set_user_pf_options(net, mode=pp_calc_mode)
     net_params["medium_temp_C"] = net_data.at[0, "TEMP"]
     net_params["medium_temp_K"] = net_data.at[0, "TEMP"] + 273.15
     net_params["calculation_results_valid"] = not bool(net_data.at[0, "CALCDIRTY"])
@@ -199,7 +200,7 @@ def get_net_params(net, stored_data):
                        " lead to incorrect results. The possible friction models are %s."
                        % (net_params["friction_model"], known_str))
     else:
-        pandapipes.set_user_pf_options(
+        set_user_pf_options(
             net, friction_model=known_friction_models[net_params["friction_model"]],
             max_iter_hyd=net_params["max_iterations"], max_iter_therm=net_params["max_iterations"]
         )
@@ -207,7 +208,7 @@ def get_net_params(net, stored_data):
         logger.warning("The compressibility model %s is not implemented in pandapipes, which might "
                        "lead to wrong results." % net_params["compress_model"])
 
-    pandapipes.set_user_pf_options(net, ambient_temperature=net_data.at[0, "AIRTEMP"] + 273.15)
+    set_user_pf_options(net, ambient_temperature=net_data.at[0, "AIRTEMP"] + 273.15)
     state = 'liquid' if net_params['medium'] == 'W' else 'gas'
     fluid = pandapipes.create_constant_fluid(
         'STANET_fluid', state, density=net_params["rho"], viscosity=net_params["eta"],
