+++ src/pandapipes/pipeflow.py
@@ -12,12 +12,12 @@ from pandapipes.pf.build_system_matrix import build_system_matrix
 from pandapipes.pf.derivative_calculation import (calculate_derivatives_hydraulic,
                                                   calculate_derivatives_thermal)
 from pandapipes.pf.pipeflow_setup import (
-    get_net_option, get_net_options, set_net_option, init_options, create_internal_results,
-    write_internal_results, get_lookup, create_lookups, initialize_pit, reduce_pit,
-    set_user_pf_options, init_all_result_tables, identify_active_nodes_branches, check_infeed_number,
-    PipeflowNotConverged
+    init_options, create_lookups, initialize_pit, reduce_pit,
+    init_all_result_tables, identify_active_nodes_branches
 )
-from pandapipes.pf.result_extraction import extract_all_results, extract_results_active_pit
+from pandapipes.utils.internals import check_infeed_number, PipeflowNotConverged, get_net_option, get_net_options, \
+    set_net_option, get_lookup, set_user_pf_options, create_internal_results, write_internal_results
+from pandapipes.utils.result_extraction import extract_all_results, extract_results_active_pit
 
 try:
     import pandaplan.core.pplog as logging
