+++ src/pandapipes/timeseries/run_time_series.py
@@ -5,12 +5,13 @@
 import tempfile
 
 from pandapower.control import NetCalculationNotConverged
-
-from pandapipes.pipeflow import PipeflowNotConverged, pipeflow
 from pandapower.control.util.diagnostic import control_diagnostic
 from pandapower.timeseries.output_writer import OutputWriter
 from pandapower.timeseries.run_time_series import init_time_series as init_time_series_pp, cleanup,\
     run_loop
+
+from pandapipes.pipeflow import pipeflow
+from pandapipes.utils.internals import PipeflowNotConverged
 
 try:
     import pandaplan.core.pplog as logging
