+++ src/pandapipes/utils/__init__.py
@@ -2,6 +2,5 @@
 # and Energy System Technology (IEE), Kassel, and University of Kassel. All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
 
-
-from pandapipes.pf.build_system_matrix import *
-from pandapipes.pf.pipeflow_setup import *
+from pandapipes.utils.internals import *
+from pandapipes.utils.result_extraction import *
